memory problem
To calculate the coherence matrix, we need to general a very big cell
containing a lot clusters, every clustter is class of
ECCEClusterCoherence.
Thus the cell became bigger and bigger after the calculation begins.
The cluster Hamiltonian, cluster coherence will be stored in the
cluster (a cluster cell element). Now, the element of the cluster cell
will be deleted after the calculation of the  corresponding cluster
coherence.