change the method to generate a superoperator from a operator
change the function circleC in the KronProd
change the position of the class: GeneralSpinInteraction and InteractionString.
add a new class SuperInteractionString to generate a Superoperator from a string
change the class InteractionString: if one want to generate a operator from a string,
he/she must assign the matrix_strategy first.